alsa_sound: Fix audio crash during secure playback with HDMI
When the audio playback is in progress, if the HDMI is frequently
connected/disconnected crash is seen.
Fix the issue by (1) avoid calling DDP endpoint parameters when
the current playback is not for AC3 or EAC3 playback. (2) validate
the error status if the get multiple values from driver fails

Change-Id: I305540ac30ace783e9e9d7020ccf7b2ffcf5bd9d
CRs-Fixed: 473859